requirements.txt
asgiref==3.5.0
cloudinary==1.29.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2
django-allauth==0.49.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-phonenumber-field==6.1.0
django-summernote==0.8.20.0
gunicorn==20.1.0
oauthlib==3.2.0
phonenumbers==8.12.44
psycopg2==2.9.3
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests-oauthlib==1.3.1
sqlparse==0.4.2
